Bruker kun tidslinje for tilkjent ytelse

+++ domenetjenester/beregning-ytelse/src/main/java/no/nav/ung/sak/ytelse/DagsatsOgUtbetalingsgrad.java
@@ -3,6 +3,6 @@ package no.nav.ung.sak.ytelse;
 
 import java.math.BigDecimal;
 
-public record DagsatsOgUtbetalingsgrad(BigDecimal dagsats, BigDecimal utbetalingsgrad) {
+public record DagsatsOgUtbetalingsgrad(Long dagsats, BigDecimal utbetalingsgrad) {
 }
